pizza_chatbot_langchain/requirements.txt
# requirements.txt
langchain-google-genai
python-dotenv
langchain
rich